Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    float a;
+    double a;
     scanf("%lf", &a);
-    printf("You entered: %.4f", a);
+    printf("You entered: %lf", a);
     return 0;
 }
